@wdio/cli: refactor
