--- packages/wdio-cli/src/@types/async-exit-hooks.d.ts
@@ -1,5 +1,0 @@
-declare module 'async-exit-hook' {
-    export default function exitHook (
-        hook: (callback: () => void) => void
-    ): void
-}
